Reduce memory usage: visit files biggest first
By experimentation, the file order significantly affects process
memory usage. The theory is that the JVM heap becomes defragmented
after lots of little allocs and when a large chunk is asked for, the
heap is expanded instead of rearranging free chunks.

The method then is to do the big files first so when they leave the
heap, there is a big open workspace.